refactor: Roster Viewer
